play_zork/requirements.txt
# Visual Game Agent for Zork - Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Transformers for Qwen3-VL (need 4.57.0+ for Qwen3-VL support)
transformers>=4.57.0
accelerate>=0.20.0

# Qwen-VL utilities for image processing
qwen-vl-utils>=0.0.14

# Screen capture and automation
mss>=9.0.0
pyautogui>=0.9.54
Pillow>=10.0.0
